// Angular/src/app/enum/sidebar-enum.ts
export enum SIDEBAR_ENUM {
  GET_IP_ADDRESS = '取得IP',
  COPY_AND_LOCK = '鎖滑鼠右鍵&複製',
  PDF_VIEWER = '預覽PDF',
  MOVING_DIV = '移動視窗'
}

export enum SIDEBAR_ROUTING {
  GET_IP_ADDRESS = 'getIpAddress',
  COPY_AND_LOCK = 'copyAndLock',
  PDF_VIEWER = 'pdfViewer',
  MOVING_DIV = 'movingDiv'
}